fix(reloc): widen ptrLoadSaveObj to uintptr_t; add Reloc64_SCVertArray
Two 64-bit pointer-width bugs fixed:

1. sdata->ptrLoadSaveObj was `int` — a 32-bit field used to store a
   native heap pointer, truncating the upper 32 bits. Changed to
   `uintptr_t` (4 bytes on 32-bit = layout preserved, 8 bytes on native).
   Write site in SelectProfile.c updated from `(int)obj` to `(uintptr_t)obj`.

2. SCVerts were relocated as a `// leaf` in Reloc64_Level, meaning
   ptrSCVert received only ctx->base + disc_offset. At runtime, each
   scVert->v was read as 8 native bytes from a 4-byte disc slot,
   producing garbage (e.g. 0x00eadb714433ff57), and scVert++ advanced
   by native sizeof(SCVert)=20 not disc stride 16, causing AnimateQuad
   to crash immediately on the first animated water vertex.

   Added Reloc64_SCVertArray (mirrors Reloc64_WaterVertArray) with disc
   layout constants for SCVert (0x10 bytes: v@0, pos_xy@4, pos_zw@8,
   color@C). Now allocates a native-sized array, resolves v via
   Reloc64_Resolve (leaf), and copies the three plain-int fields.

+++ game/SelectProfile.c
@@ -257,7 +257,7 @@ void SelectProfile_Init(u16 flags)
 		// through this allocation result before its later null check; keep
 		// unpatched until a valid menu repro proves the allocation can fail.
 		obj = (struct SelectProfileLoadSaveObj *)t->object;
-		sdata->ptrLoadSaveObj = (int)obj;
+		sdata->ptrLoadSaveObj = (uintptr_t)obj;
 		obj->icons = (struct SelectProfileLoadSaveIcon *)&sdata->LoadSaveData[0];
 		memset(obj->icons, 0, sizeof(sdata->LoadSaveData));
